+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### UNRELEASED CHANGES
 - Add: Set up e2e Cypress tests for login with FirstBook
+- Fix: Block users from navigating directly to books above their age range
 
 ### 4.5.0
 - Add: Update cypress test info in README